Add hardware type to controller_hardware_connector
Including hardware type is necessary to group the connectors
based on hardware. Only one type of connector is active per
hardware. This is only a software grouping and does not actually
check if the correct hardware is being used or not. This allows
for using optimal control connectors which use both quadrotor
and arm to be grouped under one of the existing hardware type
categories.