Add script to generate dot files for state machine
The dot files are stored .dot_out folder. The script is run
using generate_dot_files.py file which is run when cmake is used
It automatically tries to graph all the files in state machine
folder except base state machine and state machine connector.
Additional files can be added to the script