Add ability to chain action sequences with shorting
The shorting action sequences call a sequence of actions one after
another. If any action returns true in the sequence it stops executing
the ones after that. This is used when the actions return true if they
process a state machine event which possibly changes the state we are in
and we do not want to execute further actions without re-evaluating what
to do in the new state